Closes #57

+++ ClimbWeb/Climb/General/ImageRules.cs
@@ -2,13 +2,13 @@
 {
     public class ImageRules
     {
-        public int MaxSize { get; }
+        public long MaxSize { get; }
         public int Width { get; }
         public int Height { get; }
         public string Folder { get; }
         public string MissingUrl { get; }
 
-        public ImageRules(int maxSize, int width, int height, string folder, string missingUrl)
+        public ImageRules(long maxSize, int width, int height, string folder, string missingUrl)
         {
             MaxSize = maxSize;
             Width = width;
